requirements.txt
# OKX Smart Trading Agent - Python Dependencies

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
web3>=6.11.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# API and networking
aiohttp>=3.9.0
httpx>=0.25.0

# Cryptography and security
cryptography>=41.0.0
pyjwt>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging and monitoring
loguru>=0.7.0

# Configuration
pyyaml>=6.0.0

# Utilities
python-dateutil>=2.8.2
